+++ test/td_dd/xlsx/reader_test.exs
@@ -93,6 +93,7 @@ defmodule TdDd.XLSX.ReaderTest do
         %{
           "cardinality" => "?",
           "name" => "hierarchy_name_1",
+          "label" => "Hierarchy name 1",
           "type" => "hierarchy",
           "values" => %{"hierarchy" => %{"id" => 1}},
           "widget" => "dropdown"
@@ -100,6 +101,7 @@ defmodule TdDd.XLSX.ReaderTest do
         %{
           "cardinality" => "*",
           "name" => "hierarchy_name_2",
+          "label" => "Hierarchy name 2",
           "type" => "hierarchy",
           "values" => %{"hierarchy" => %{"id" => 1}},
           "widget" => "dropdown"