+++ test/td_dq_web/controllers/rule_upload_controller_test.exs
@@ -10,6 +10,7 @@ defmodule TdDqWeb.RuleUploadControllerTest do
         %{
           "cardinality" => "?",
           "name" => "hierarchy_name_1",
+          "label" => "Hierarchy name 1",
           "type" => "hierarchy",
           "values" => %{"hierarchy" => %{"id" => 1}},
           "widget" => "dropdown"
@@ -17,6 +18,7 @@ defmodule TdDqWeb.RuleUploadControllerTest do
         %{
           "cardinality" => "*",
           "name" => "hierarchy_name_2",
+          "label" => "Hierarchy name 2",
           "type" => "hierarchy",
           "values" => %{"hierarchy" => %{"id" => 1}},
           "widget" => "dropdown"